refactor(effects/v3): IdList is segments, and each segment writes itself
A list stops being one representation that degrades and becomes a sequence
of segments. A push that continues the current segment extends it; one that
does not opens a new segment. Nothing collapses the whole list, and nothing
is rebuilt at encode time to discover its shape.

Two variants, and each one encodes itself, so the logical enum in
`id_list.rs` and the wire enum in `mod.rs` are now the same thing:

* `Range { base, len }` — consecutive ids. A single id is a range of one,
  which is why this alone describes any list however unordered, and why
  `Plain` and `Dictionary` are gone.
* `Ascending { bitmap }` — what several ranges collapse into when they
  ascend and `serialized_size()` says the bitmap is genuinely cheaper.
  Exact rather than estimated, so a second implementation reaches the same
  answer.

The wire carries no segment count. The record already says how many ids it
holds, so the reader takes segments until that is satisfied — which lets the
final segment leave its length implied. Writing the count instead cost four
bytes per record and measured +14.6% on a payload of ten thousand single-id
records; without it a lone id is two bytes, where the encoding this replaces
charged three.

Measured against the ladder it replaces, same shapes, same test:

| shape                    |     before |      after |
|--------------------------|-----------:|-----------:|
| one shape, 10k nodes     |    120,023 |    120,022 |
| 20 shapes                |    140,402 |    141,022 |
| 500 shapes               |    150,002 |    168,246 |
| 2,000 shapes             |    180,002 |    193,746 |
| 10,000 singleton records |    339,746 |    329,746 |
| alternating labels       |    140,042 |    136,488 |
| 10,000 supernode edges   |     10,058 |     60,019 |

The last row is the cost of dropping the dictionary and is not yet paid for
— see the follow-up commit.

+++ graph/src/effects/error.rs
@@ -95,7 +95,7 @@ pub enum DecodeError {
 
     /// A `Range` block whose `base + count` would wrap past `u64`.
     #[error("id range starting at {base} cannot hold {count} ids")]
-    BadRange { base: u64, count: u32 },
+    BadRange { base: u64, count: u64 },
 
     /// More ids in one record than any real query produces.
     ///
